Fixed Script references being out of order Fixed Packages not having the second PLD2 entry Overhauled T4 generation on records Some work on T4 SubrecordCollection generation Implemented some rudimentary sorting

+++ ESPSharp/Enums/Flags/AmbushPackageFlags.cs
@@ -7,7 +7,7 @@ using System.Threading.Tasks;
 namespace ESPSharp.Enums.Flags
 {
     [Flags]
-    public enum AmbushPackageFlags : uint
+    public enum AmbushPackageFlags : ushort
     {
         HideWhileAmbushing = 0x01
     }
